(dbatools) Updates dbatools to use dependency on dbatools.library
Dbatools now has a library it is dependent on. This was not reflected in the package.

This update to the package and update logic adds a dependency on the dbatools-library.powershell package, and ensures that on a new version of the dbatools package it should be pointing at the right new version of that package.